requirements.txt
pandas
python-dotenv
langchain
openai
faiss-cpu
tiktoken
datefinder
langchain_community
rank_bm25
streamlit-cookies-manager
